# latest_openapi.json
{"openapi":"3.1.0","info":{"title":"WeRSS API","description":"微信公众号RSS生成服务API文档","version":"1.0.0"},"paths":{"/api/v1/wx/auth/login":{"post":{"tags":["认证"],"summary":"用户登录","operationId":"login_api_v1_wx_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_v1_wx_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/wx/auth/logout":{"post":{"tags":["认证"],"summary":"用户注销","operationId":"logout_api_v1_wx_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/wx/auth/refresh":{"post":{"tags":["认证"],"summary":"刷新Token","operationId":"refresh_token_api_v1_wx_auth_refresh_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/wx/auth/verify":{"get":{"tags":["认证"],"summary":"验证Token有效性","description":"验证当前token是否有效","operationId":"verify_token_api_v1_wx_auth_verify_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/wx/user":{"get":{"tags":["用户管理"],"summary":"获取用户信息","operationId":"get_user_info_api_v1_wx_user_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"put":{"tags":["用户管理"],"summary":"修改用户资料","description":"修改用户基本信息(不包括密码)","operationId":"update_user_info_api_v1_wx_user_put","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Update Data"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/wx/user/password":{"put":{"tags":["用户管理"],"summary":"修改密码","description":"修改用户密码","operationId":"change_password_api_v1_wx_user_password_put","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Password Data"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/wx/user/avatar":{"post":{"tags":["用户管理"],"summary":"上传用户头像","description":"处理用户头像上传","operationId":"upload_avatar_api_v1_wx_user_avatar_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_avatar_api_v1_wx_user_avatar_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/wx/articles":{"get":{"tags":["文章管理"],"summary":"获取文章列表","operationId":"get_articles_api_v1_wx_articles_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":5,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/wx/articles/{article_id}":{"get":{"tags":["文章管理"],"summary":"获取文章详情","operationId":"get_article_detail_api_v1_wx_articles__article_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"article_id","in":"path","required":true,"schema":{"type":"integer","title":"Article Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/wx/mps":{"get":{"tags":["公众号管理"],"summary":"获取公众号列表","operationId":"get_mps_api_v1_wx_mps_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/wx/mps/{mp_id}":{"get":{"tags":["公众号管理"],"summary":"获取公众号详情","operationId":"get_mp_detail_api_v1_wx_mps__mp_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"mp_id","in":"path","required":true,"schema":{"type":"string","title":"Mp Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/{path}":{"get":{"summary":"Serve Vue App","description":"处理Vue应用路由","operationId":"serve_vue_app__path__get","parameters":[{"name":"path","in":"path","required":true,"schema":{"type":"string","title":"Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Serve Root","description":"处理根路由","operationId":"serve_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_login_api_v1_wx_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_api_v1_wx_auth_login_post"},"Body_upload_avatar_api_v1_wx_user_avatar_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_avatar_api_v1_wx_user_avatar_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}}}